Refactor configuration service
