+++ app/services/configuration/env_configuration_service.rb
@@ -1,0 +1,13 @@
+module Configuration
+  class EnvConfigurationService < ConfigurationService
+  private
+
+    def config_present?
+      raise NotImplementedError
+    end
+
+    def read_config
+      raise NotImplementedError
+    end
+  end
+end
