+++ app/services/configuration/paas_configuration_service.rb
@@ -1,28 +1,12 @@
 module Configuration
   class PaasConfigurationService < ConfigurationService
-    def initialize(logger = Rails.logger)
-      super()
-      @logger = logger
-      @paas_config = read_pass_config
-      @s3_buckets = read_s3_buckets
-      @redis_uris = read_redis_uris
-    end
-
-    def s3_config_present?
-      config_present? && @paas_config.key?(:"aws-s3-bucket")
-    end
-
-    def redis_config_present?
-      config_present? && @paas_config.key?(:redis)
-    end
-
   private
 
     def config_present?
       !ENV["VCAP_SERVICES"].nil?
     end
 
-    def read_pass_config
+    def read_config
       unless config_present?
         @logger.warn("Could not find VCAP_SERVICES in the environment!")
         return {}
@@ -34,30 +18,6 @@ module Configuration
         @logger.warn("Could not parse VCAP_SERVICES!")
         {}
       end
-    end
-
-    def read_s3_buckets
-      return {} unless s3_config_present?
-
-      s3_buckets = {}
-      @paas_config[:"aws-s3-bucket"].each do |bucket_config|
-        if bucket_config.key?(:instance_name)
-          s3_buckets[bucket_config[:instance_name].to_sym] = bucket_config
-        end
-      end
-      s3_buckets
-    end
-
-    def read_redis_uris
-      return {} unless redis_config_present?
-
-      redis_uris = {}
-      @paas_config[:redis].each do |redis_config|
-        if redis_config.key?(:instance_name)
-          redis_uris[redis_config[:instance_name].to_sym] = redis_config.dig(:credentials, :uri)
-        end
-      end
-      redis_uris
     end
   end
 end
