+++ spec/services/configuration/env_configuration_service_spec.rb
@@ -1,0 +1,17 @@
+require "rails_helper"
+
+Rspec.describe Configuration::EnvConfigurationService do
+  before do
+    # Do nothing
+  end
+
+  after do
+    # Do nothing
+  end
+
+  context 'when condition' do
+    it 'succeeds' do
+      pending 'Not implemented'
+    end
+  end
+end
